refactor: mv validateRayJobSpec func and unit test to validation.go
